Move to to_value.go
